+++ common/src/main/java/com/sigmapool/common/models/EarningsDto.kt
@@ -1,5 +1,5 @@
 package com.sigmapool.common.models
 
 class EarningsDto(
-    val earnings: Int
+    val earnings: Float
 )
